+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Maturity moved to kscience extension.
 - Readme extension now requires kscience extension.
 - Project plugin overhaul (not only publish)
+- ABI validation is configurable from the root project
 
 ### Deprecated
 
@@ -21,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Context parameter flag
+- Deploy problem with opensavvy-resources
 
 ### Security